+++ packages/atomic/src/components/ipx/atomic-ipx-modal/atomic-ipx-modal.tsx
@@ -33,11 +33,13 @@ export class AtomicIPXModal implements InitializableComponent<AnyBindings> {
   @State() public error!: Error;
 
   @Prop({mutable: true}) source?: HTMLElement;
+
   /**
    * The container to hide from the tabindex and accessibility DOM when the modal is closed.
    */
   @Prop({mutable: true}) container?: HTMLElement;
   @Prop({reflect: true, mutable: true}) isOpen = false;
+  @Prop({reflect: true}) noFocusTrap = false;
 
   @Event() animationEnded!: EventEmitter<never>;
 
@@ -51,16 +53,14 @@ export class AtomicIPXModal implements InitializableComponent<AnyBindings> {
 
     if (isOpen) {
       document.body.classList.add(modalOpenedClass);
-      if (watchToggleOpenExecution !== this.currentWatchToggleOpenExecution) {
-        return;
+      if (watchToggleOpenExecution === this.currentWatchToggleOpenExecution) {
+        !this.noFocusTrap && (this.focusTrap!.active = true);
       }
-      this.focusTrap!.active = true;
-    } else {
-      document.body.classList.remove(modalOpenedClass);
-      if (watchToggleOpenExecution !== this.currentWatchToggleOpenExecution) {
-        return;
-      }
-      this.focusTrap!.active = false;
+      return;
+    }
+    document.body.classList.remove(modalOpenedClass);
+    if (watchToggleOpenExecution === this.currentWatchToggleOpenExecution) {
+      !this.noFocusTrap && (this.focusTrap!.active = false);
     }
   }
 
@@ -91,58 +91,30 @@ export class AtomicIPXModal implements InitializableComponent<AnyBindings> {
   public render() {
     this.updateBreakpoints();
 
+    const Body = () => (
+      <atomic-ipx-body isOpen={this.isOpen}>
+        <slot name="header" slot="header" />
+        <slot name="body" slot="body" />
+        <slot name="footer" slot="footer" />
+      </atomic-ipx-body>
+    );
+
     return (
       <Host class={this.getClasses().join(' ')}>
-        <div
-          part="backdrop"
-          class="fixed left-0 top-0 right-0 bottom-0 z-[9999]"
-        >
-          <atomic-focus-trap
-            role="dialog"
-            aria-modal={this.isOpen.toString()}
-            source={this.source}
-            container={this.container ?? this.host}
-            ref={(ref) => (this.focusTrap = ref)}
-          >
-            <article
-              part="container"
-              class={`${this.isOpen ? 'visible' : 'invisible'}`}
-              onAnimationEnd={() => this.animationEnded.emit()}
+        <div part="backdrop">
+          {this.noFocusTrap ? (
+            <Body />
+          ) : (
+            <atomic-focus-trap
+              role="dialog"
+              aria-modal={this.isOpen.toString()}
+              source={this.source}
+              container={this.container ?? this.host}
+              ref={(ref) => (this.focusTrap = ref)}
             >
-              <header part="header-wrapper" class="flex flex-col items-center">
-                <div part="header">
-                  <slot name="header"></slot>
-                </div>
-              </header>
-              <hr part="header-ruler" class="border-neutral"></hr>
-              <div
-                part="body-wrapper"
-                class="overflow-auto grow flex flex-col w-full"
-              >
-                <div
-                  part="body"
-                  class="w-full max-w-lg"
-                  ref={(element) =>
-                    element?.addEventListener(
-                      'touchmove',
-                      (e) => this.isOpen && e.stopPropagation(),
-                      {passive: false}
-                    )
-                  }
-                >
-                  <slot name="body"></slot>
-                </div>
-              </div>
-              <footer
-                part="footer-wrapper"
-                class="border-neutral border-t bg-background z-10 flex flex-col items-center w-full"
-              >
-                <div part="footer" class="max-w-lg">
-                  <slot name="footer"></slot>
-                </div>
-              </footer>
-            </article>
-          </atomic-focus-trap>
+              <Body />
+            </atomic-focus-trap>
+          )}
         </div>
       </Host>
     );