+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.9.10](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.9.9...@coveo/quantic@2.9.10) (2023-02-06)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.9.9](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.9.8...@coveo/quantic@2.9.9) (2023-02-02)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.9.8](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.9.7...@coveo/quantic@2.9.8) (2023-02-01)
+
+**Note:** Version bump only for package @coveo/quantic
+
 ## [2.9.7](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.9.6...@coveo/quantic@2.9.7) (2023-01-25)
 
 ### Bug Fixes